Added base Router functionality and reworked Model structure

+++ Core/Model.php
@@ -11,42 +11,55 @@ class Model implements IModel
 {
     protected $db;
     protected $tableName;
+    protected $bootstrap;
     //protected $fields = [];
 
-    public function __construct(IDatabase $db, string $tableName)//, array $fields)
+    public function __construct()
     {
-        $this->db = $db;
-        $this->tableName=$tableName;
-        //$this->fields=$fields;
+        $this->bootstrap = \Bootstrap::getInstance();
+        $this->db = $this->bootstrap->db;
     }
 
+    public function findTable(string $tableName)
+    {
+        $this->tableName=$tableName;
+    }
+
+    public function validate()
+    {
+        if(!isset($this->tableName)) {
+            $this->bootstrap->getLogger()->logError(41);
+            return false;
+        }
+        else return true;
+    }
     public function getName() : string
     {
-        return $this->tableName;
+        return ($this->validate()) ? $this->tableName : null;
     }
 
     public function insert(array $params) :PDOStatement
     {
-        return $this->db->insert($this->tableName,$params);
+        if($this->validate()) return $this->db->insert($this->tableName,$params);
     }
     public function update(int $id, array $params) :PDOStatement
     {
-        return $this->db->update($this->tableName,$id,$params);
+        if($this->validate()) return $this->db->update($this->tableName,$id,$params);
     }
     public function delete(int $id) :PDOStatement
     {
-        return $this->db->delete($this->tableName,$id);
+        if($this->validate()) return $this->db->delete($this->tableName,$id);
     }
     public function get(int $id) :PDOStatement
     {
-        return $this->db->get($this->tableName,$id);
+        if($this->validate()) return $this->db->get($this->tableName,$id);
     }
     public function getAll() :PDOStatement
     {
-       return $this->db->getAll($this->tableName);
+        if($this->validate()) return $this->db->getAll($this->tableName);
     }
     public function query(string $query) :PDOStatement
     {
-        return $this->db->query($query);
+        if($this->validate()) return $this->db->query($query);
     }
 }
